QA_Env_Auto_Test/requirements.txt
pytest==8.3.3
requests==2.32.3
python-socketio[client]==5.11.3
allure-pytest==2.13.5


